sua theo db moi

+++ src/main/java/com/example/internhub_be/payload/request/InternshipPositionRequest.java
@@ -18,5 +18,5 @@ public class InternshipPositionRequest {
     @Size(max = 1000, message = "Description must not exceed 1000 characters")
     private String description;
 
-    private Integer departmentId;
+    private Long departmentId;
 }
