+++ src/main/java/com/example/internhub_be/repository/InternshipPositionRepository.java
@@ -3,5 +3,5 @@ package com.example.internhub_be.repository;
 import com.example.internhub_be.domain.InternshipPosition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InternshipPositionRepository extends JpaRepository<InternshipPosition, Integer> {
+public interface InternshipPositionRepository extends JpaRepository<InternshipPosition, Long> {
 }
